Move selected id to topper level
